Exchange notifies processor of payments
The current proposal has the processor independently determine when payment is made.   This adds significant complexity to each processor and creates the possibility that the processor thinks a payment was made when the exchange does not (or vice versa).

Having the exchange notify the processor of payments via a webhook solves this.   This is easier for the exchange to implement since they are already crypto experts, lowers the barrier for processor adoption (since they don't now need to be crypto experts), and ensures that invoices are marked as paid only when they really are.